-Removed some debug code. Left some intact (there is no GUI info display yet for connection list, for example). -Fixed connection timeout. It's really fast now. -Adjusted formatting.

+++ trunk/src/scripts/DefaultProperties.java
@@ -46,7 +46,7 @@ public class DefaultProperties {
 		
 		/* network package properties */
 		props.setProperty("network::port", new Integer(24600)); /* listening port */
-		props.setProperty("network::timeout", new Integer(4000)); /* connection timeout */
+		props.setProperty("network::timeout", new Integer(2000)); /* connection timeout */
 		props.setProperty("network::so_timeout", new Integer(35)); /* network timeout */
 		
 		/* operations package properties */
